+++ src/services/dashboards/dashboard-detail/lib/__tests__/index.test.ts
@@ -5,7 +5,7 @@ import {
     WIDGET_CONTAINER_MIN_WIDTH,
 } from '@/services/dashboards/dashboard-detail/lib/config';
 import type { WidgetSize } from '@/services/dashboards/dashboard-detail/lib/type';
-import { widgetWidthAssigner } from '@/services/dashboards/dashboard-detail/lib/widthHelper';
+import { widgetWidthAssigner } from '@/services/dashboards/dashboard-detail/lib/width-helper';
 
 
 /**
