+++ src/services/dashboards/dashboard-detail/lib/width-helper.ts
